strlcat and strlcopy stuff, fixed prototypes, need to test with stddef included in libft.h

+++ src/ft_bzero.c
@@ -1,6 +1,6 @@
-void	ft_bzero(void *s, int n)
+void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	unsigned int	i;
 	unsigned char	*str;
 
 	i = 0;
