+++ src/ft_memchr.c
@@ -1,4 +1,4 @@
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned int	i;
 	const unsigned char	*str;
