+++ src/ft_memcmp.c
@@ -1,4 +1,4 @@
-int	ft_memcmp(const void *s1, const void *s2, unsigned int n)
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned int	i;
 	unsigned const char	*str_1;
