+++ src/ft_memset.c
@@ -1,4 +1,4 @@
-void	*ft_memset(void *s, int c, int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	int	i;
 	char	*str;
